Fix #26 and improve !award_ceremony
In addition to fixing the bug, this adds the option to add an argument
to !award_ceremony that specifys the day-offset (how many days before
current date) of the award ceremony. This defaults to 0 (the current date)

The award ceremony is still automatically executed at midnight, which is good.
(way better than having it @23:59 and hoping it takes <1min to process :DD)

(This commit also includes some formatting => `cargo fmt`)