Add a button to voting that does nothing.
The intentions behind this commit is to prevent the "real votes" to be
abused for "autistic-reasons". This adds an additionnal button that is
there for individuals that can't stop clicking the buttons
(I admit, they are quite difficult to resist)

This is a feature that was requested by some members of the council.